agent/requirements.txt
camel-ai
openai
pydantic
requests
flask
flask-cors
pyyaml
python-dotenv
